+++ crates/sase_core_py/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- bind `validate_finalizer_plan` and `authenticate_finalizer_plan` for sealed-plan integrity
+
 ## [0.29.6](https://github.com/sase-org/sase-core/compare/v0.29.5...v0.29.6) - 2026-08-21
 
 ### Added